small corrections
COMPTS is already initialized with READBMP. COMPTD has to be
initialized instead. Otherwise a small shift appears in the image. This
was not visible on the original demo picture because of the surrounding
black background.
Removed one unnecessary assignment to BUFFDST since it was already
written on screen and will be overwritten thereafter.