Replaced hi-res address calculation
Replaced hi-res address calculation with the optimized new Woz
calculation <http://www.txbobsc.com/aal/1986/aal8612.html#a9>
Swapped the order of Y and X register initialisation in SAVEDST because
Y is used in the Woz routine.
Added the image of Jonathan Mak in the ProDOS Disk image and the
samples folder.